refactor(reports-rest): refactoring of the OpenAPI generation
This refacoring is a preparation to include
