+++ reports-rest/src/main/java/org/jboss/da/rest/api/v1/Reports.java
@@ -1,4 +1,4 @@
-package org.jboss.da.rest.reports;
+package org.jboss.da.rest.api.v1;
 
 import io.opentelemetry.instrumentation.annotations.SpanAttribute;
 import io.opentelemetry.instrumentation.annotations.WithSpan;
